+++ app/Repositories/OptionRepository.php
@@ -154,4 +154,35 @@ class OptionRepository
 
         return $option;
     }
+
+    /**
+     * Updates a service option's scripts in the database.
+     *
+     * @param  int    $id
+     * @param  array  $data
+     * @return \Pterodactyl\Models\ServiceOption
+     *
+     * @throws \Pterodactyl\Exceptions\DisplayValidationException
+     */
+    public function scripts($id, array $data)
+    {
+        $option = ServiceOption::findOrFail($id);
+
+        $data['script_install'] = empty($data['script_install']) ? null : $data['script_install'];
+
+        $validator = Validator::make($data, [
+            'script_install' => 'sometimes|nullable|string',
+            'script_is_privileged' => 'sometimes|required|boolean',
+            'script_entry' => 'sometimes|required|string',
+            'script_container' => 'sometimes|required|string',
+        ]);
+
+        if ($validator->fails()) {
+            throw new DisplayValidationException(json_encode($validator->errors()));
+        }
+
+        $option->fill($data)->save();
+
+        return $option;
+    }
 }