+++ routes/daemon.php
@@ -25,3 +25,4 @@ Route::get('/services', 'ServiceController@list')->name('daemon.services');
 Route::get('/services/pull/{service}/{file}', 'ServiceController@pull')->name('daemon.pull');
 Route::get('/packs/pull/{uuid}', 'PackController@pull')->name('daemon.pack.pull');
 Route::get('/packs/pull/{uuid}/hash', 'PackController@hash')->name('daemon.pack.hash');
+Route::get('/details/option/{server}', 'OptionController@details')->name('daemon.option.details');